new commit

+++ app/Http/Controllers/AGestionCreditController.php
@@ -27,6 +27,13 @@ class AGestionCreditController extends Controller
     }
 
 
+    public function  CreditDecaisseHomePage()
+    {
+        return view("gestion_credit.pages.credit-decaisse");
+    }
+
+
+
     public function store(Request $request)
     {
         $validator = validator::make($request->all(), [
@@ -90,7 +97,9 @@ class AGestionCreditController extends Controller
             ]);
 
             foreach ($request->file('images') as $image) {
-                $path = $image->store('credits', 'public'); // Stocke dans storage/app/public/credits
+                // Conserver le nom original mais ajouter un timestamp devant
+                $filename = date('Ymd_His') . '_' . $image->getClientOriginalName();
+                $path = $image->storeAs('credits', $filename, 'public'); // Stocke dans storage/app/public/credits
                 $credit->images()->create([
                     'path' => $path
                 ]);
@@ -114,7 +123,11 @@ class AGestionCreditController extends Controller
 
     public function getCreditValidation()
     {
-        $credits = DB::table('credits')->limit(10)->get();
+        $credits = DB::table('credits')
+            ->where("statutDossier", "!=", "Décaissé")
+            ->orderBy('id_credit', 'desc') // tri décroissant sur la colonne id
+            ->limit(10)
+            ->get();
 
         foreach ($credits as $credit) {
             $credit->images = DB::table('credits_images')
@@ -158,13 +171,47 @@ class AGestionCreditController extends Controller
         ]);
     }
 
-    public function getCreditToDelete($id)
+
+    public function getSearchedCreditDecaisse($ref)
     {
-        Credits::where("id_credit", $id)->delete();
+
+        $credits = DB::table('credits')
+            ->where(function ($query) use ($ref) {
+                $query->where('NumCompte', $ref)
+                    ->orWhere('NomCompte', 'LIKE', '%' . $ref . '%');
+            })
+            ->where("statutDossier", "Décaissé")
+            ->orderBy('id_credit', 'desc') // tri décroissant sur la colonne id
+            ->limit(10)
+            ->get();
+
+
+        // Ajout des images pour chaque crédit trouvé
+        foreach ($credits as $credit) {
+            $credit->images = DB::table('credits_images')
+                ->where('credits_id', $credit->id_credit)
+                ->pluck('path'); // Retourne un tableau simple
+        }
         return response()->json([
             "status" => 1,
-            "msg" => "Dossier de crédit supprimé avec succès"
+            "data" => $credits
         ]);
+    }
+
+    public function getCreditToDelete($id)
+    {
+        if (Auth::user()->role == "DG") {
+            Credits::where("id_credit", $id)->delete();
+            return response()->json([
+                "status" => 1,
+                "msg" => "Dossier de crédit supprimé avec succès"
+            ]);
+        } else {
+            return response()->json([
+                "status" => 0,
+                "msg" => "Vous n'avez pas une autorisation requise pour supprimer un dossier de crédit"
+            ]);
+        }
     }
 
 
@@ -227,37 +274,46 @@ class AGestionCreditController extends Controller
     public function updateDossier(Request $request)
     {
 
-        Credits::where("id_credit", $request->idDossier)->update([
-            "NumCompte" => $request->NumCompte,
-            "NomCompte" => $request->NomCompte,
-            "produit_credit" => $request->produit_credit,
-            "type_credit" => $request->type_credit,
-            "recouvreur" => $request->recouvreur,
-            "montant_demande" => $request->montant_demande,
-            "date_demande" => $request->date_demande,
-            "frequence_mensualite" => $request->frequence_mensualite,
-            "nombre_echeance" => $request->nombre_echeance,
-            "NumDossier" => $request->NumDossier,
-            "gestionnaire" => $request->gestionnaire,
-            "source_fond" => $request->source_fond,
-            "monnaie" => $request->monnaie,
-            "duree_credit" => $request->duree_credit,
-            "intervale_jrs" => $request->intervale_jrs,
-            "taux_interet" => $request->taux_interet,
-            "type_garantie" => $request->type_garantie,
-            "valeur_comptable" => $request->valeur_comptable,
-            "num_titre" => $request->num_titre,
-            "valeur_garantie" => $request->valeur_garantie,
-            "date_sortie_titre" => $request->date_sortie_titre,
-            "date_expiration_titre" => $request->date_expiration_titre,
-            "description_titre" => $request->description_titre,
+        $checkStatus = Credits::where("id_credit", $request->idDossier)->first();
 
-        ]);
+        if ($checkStatus->statutDossier == "Décaissé") {
+            return response()->json([
+                "status" => 0,
+                "msg" => "Impossible de modifier un dossier déjà décaissé ! "
+            ]);
+        } else {
+            Credits::where("id_credit", $request->idDossier)->update([
+                "NumCompte" => $request->NumCompte,
+                "NomCompte" => $request->NomCompte,
+                "produit_credit" => $request->produit_credit,
+                "type_credit" => $request->type_credit,
+                "recouvreur" => $request->recouvreur,
+                "montant_demande" => $request->montant_demande,
+                "date_demande" => $request->date_demande,
+                "frequence_mensualite" => $request->frequence_mensualite,
+                "nombre_echeance" => $request->nombre_echeance,
+                "NumDossier" => $request->NumDossier,
+                "gestionnaire" => $request->gestionnaire,
+                "source_fond" => $request->source_fond,
+                "monnaie" => $request->monnaie,
+                "duree_credit" => $request->duree_credit,
+                "intervale_jrs" => $request->intervale_jrs,
+                "taux_interet" => $request->taux_interet,
+                "type_garantie" => $request->type_garantie,
+                "valeur_comptable" => $request->valeur_comptable,
+                "num_titre" => $request->num_titre,
+                "valeur_garantie" => $request->valeur_garantie,
+                "date_sortie_titre" => $request->date_sortie_titre,
+                "date_expiration_titre" => $request->date_expiration_titre,
+                "description_titre" => $request->description_titre,
+                "statutDossier" => $request->statutDossier
+            ]);
 
-        return response()->json([
-            "status" => 1,
-            "msg" => "Mise à jour effectuée avec succès ! "
-        ]);
+            return response()->json([
+                "status" => 1,
+                "msg" => "Mise à jour effectuée avec succès ! "
+            ]);
+        }
     }
 
     //PERMET D'AJOUTER LE FICHIER DE SIGNATURE AU DOSSIER
@@ -270,12 +326,12 @@ class AGestionCreditController extends Controller
         ]);
 
         $credit = Credits::findOrFail($request->idDossier);
-
+        // $credit = Credits::where("id_credit", $request->idDossier)->first();
         $path = $request->file('signature_file')->store('signatures', 'public');
 
         $credit->signatures()->create([
             'signature_file' => $path,
-            'signed_by' => Auth::user()->name,
+            'signed_by' => Auth::user()->role,
         ]);
 
         return response()->json([
@@ -283,5 +339,102 @@ class AGestionCreditController extends Controller
             'msg' => 'Signature ajoutée avec succès.',
             'signature_file' => $path
         ]);
+    }
+
+    public function showTimeLine($creditId)
+    {
+        $signatures = DB::table('signatures')
+            ->where('credit_id', $creditId)
+            ->orderBy('created_at', 'asc')
+            ->get();
+
+        $timeline = [];
+        $previousDate = null;
+
+        foreach ($signatures as $sig) {
+            $delay = null;
+            if ($previousDate) {
+                $delay = \Carbon\Carbon::parse($previousDate)->diffInDays($sig->created_at);
+            }
+
+            $timeline[] = [
+                'signed_by' => $sig->signed_by,
+                'signature_file' => $sig->signature_file,
+                'signed_at' => $sig->created_at,
+                'delay_from_previous' => $delay
+            ];
+
+            $previousDate = $sig->created_at;
+        }
+
+        return response()->json($timeline);
+    }
+
+    public function getCreditDecaisse()
+    {
+        $credits = DB::table('credits')->where('statutDossier', 'Décaissé')->limit(10)->get();
+
+        foreach ($credits as $credit) {
+            $credit->images = DB::table('credits_images')
+                ->where('credits_id', $credit->id_credit)
+                ->pluck('path'); // retourne un tableau
+        }
+        return response()->json([
+            "status" => 1,
+            "data" => $credits
+        ]);
+    }
+    //PERMET D'AJOUTER UNE NOUVEAU FICHIER AU DOSSIER 
+    public function addNewFile(Request $request)
+    {
+
+        // Vérifier l'extension
+        $file = $request->file('newFile');
+
+        $allowedExtensions = ['pdf', 'jpg', 'jpeg', 'png'];
+        $extension = strtolower($file->getClientOriginalExtension());
+
+        if (!in_array($extension, $allowedExtensions)) {
+            return response()->json([
+                "status" => 0,
+                "msg" => "Format de fichier non autorisé. Seuls les PDF et les images (jpg, jpeg, png) sont acceptés."
+            ]);
+        }
+
+        $checkStatus = Credits::where("id_credit", $request->creditId)->first();
+
+        if ($checkStatus->statutDossier == "Décaissé") {
+            return response()->json([
+                "status" => 0,
+                "msg" => "Impossible de modifier un dossier déjà décaissé ! "
+            ]);
+        } else {
+
+
+            if ($request->hasFile('newFile')) {
+                $file = $request->file('newFile');
+
+                // Conserver le nom original mais ajouter un timestamp devant
+                $filename = date('Ymd_His') . '_' . $file->getClientOriginalName();
+
+                // Sauvegarder le fichier
+                $path = $file->storeAs('credits', $filename, 'public');
+
+                $checkStatus->images()->create([
+                    'path' => $path
+                ]);
+
+                return response()->json([
+                    "status" => 1,
+                    "msg" => "Nouveau fichier ajouté avec succès",
+                    "path" => $path
+                ]);
+            }
+
+            return response()->json([
+                "status" => 0,
+                "msg" => "Aucun fichier reçu"
+            ]);
+        }
     }
 }
